+++ packages/ckeditor5-ui/src/menubar/menubarmenulistitemview.ts
@@ -39,8 +39,6 @@ export default class MenuBarMenuListItemView extends ListItemView {
 			}
 		} );
 
-		// this.parentMenuView = parentMenuView;
-
 		this.delegate( 'mouseenter' ).to( parentMenuView );
 	}
 }
